Tables - Extract SingleLocation class to dedicated file
